Patch for Zoo
tracks internal commit 4ff1771495572684c2b08701e2bf3f2517f45a50

+++ plugins/Zoo/zoo.pl
@@ -310,6 +310,10 @@ sub action {
 		print getData("over_socialized");
 		return;
 	} else {
+		if ($zoo->count($form->{uid}) > $constants->{people_max}) {
+			print getData("over_socialized");
+			return 0;
+		}
 		if ( $form->{op} eq 'delete' || $form->{type} eq 'neutral') {
 			if ($form->{uid}) {
 				$zoo->delete($user->{uid}, $form->{uid});
